Add example scripts.

+++ pyffs/util.py
@@ -367,9 +367,9 @@ def ffsn_sample(T, N_FS, T_c, N_s, mod=None):
     idx = []
     for d in range(D):
         # get values for d-dimension
-        # cast to int in case cupy array is passed
+        # cast to scalar in case cupy array is passed
         _sample_points, _idx = ffs_sample(
-            T=int(T[d]), N_FS=int(N_FS[d]), T_c=int(T_c[d]), N_s=int(N_s[d]), mod=mod
+            T=float(T[d]), N_FS=int(N_FS[d]), T_c=float(T_c[d]), N_s=int(N_s[d]), mod=mod
         )
 
         # reshape for sparse array
